+++ src/Schema/SqlToolsInterface.php
@@ -2,7 +2,7 @@
 /**
  * @author     Laurent Jouanneau
  * @contributor Laurent Jouanneau
- * @copyright  2020 Laurent Jouanneau
+ * @copyright  2020-2025 Laurent Jouanneau
  *
  * @see        https://jelix.org
  * @licence     http://www.gnu.org/licenses/lgpl.html GNU Lesser General Public Licence, see LICENCE file
@@ -19,7 +19,7 @@ interface SqlToolsInterface
     /**
      * @param ConnectionInterface $connector the connection to a database
      */
-    public function __construct(ConnectionInterface $connector = null);
+    public function __construct(?ConnectionInterface $connector = null);
 
     /**
      * @return ConnectionInterface